requirements.txt
anthropic>=0.40.0
coinbase-agentkit
flask>=3.0.0
python-dotenv>=1.0.0
